Refactor: Move snap sync tests to separate file
Moves the tests for the SnapSync stage into their own file, `snap_sync_tests.rs`, to improve organization and maintainability.
